requirement.txt
streamlit==1.17.0
IMDbPY==2023.7.5
sentence-transformers==2.2.2
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.3